// src/generated/model/createEnrollmentRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CreateEnrollmentRequestEnrollmentType } from './createEnrollmentRequestEnrollmentType.ts';

export interface CreateEnrollmentRequest {
  studentId: string;
  academicYearId: string;
  intakeId: string;
  batchId: string;
  programId: string;
  programLevelId: string;
  classId?: string;
  enrollmentType: CreateEnrollmentRequestEnrollmentType;
  enrollmentDate: string;
  remarks?: string;
}