// src/generated/model/updateEnrollmentRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UpdateEnrollmentRequestEnrollmentType } from './updateEnrollmentRequestEnrollmentType.ts';
import type { UpdateEnrollmentRequestStatus } from './updateEnrollmentRequestStatus.ts';

export interface UpdateEnrollmentRequest {
  classId?: string;
  enrollmentType?: UpdateEnrollmentRequestEnrollmentType;
  status: UpdateEnrollmentRequestStatus;
  enrollmentDate: string;
  remarks?: string;
}