// src/generated/model/createEnrollmentRequestEnrollmentType.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type CreateEnrollmentRequestEnrollmentType = typeof CreateEnrollmentRequestEnrollmentType[keyof typeof CreateEnrollmentRequestEnrollmentType];


export const CreateEnrollmentRequestEnrollmentType = {
  SELF_ENROLL: 'SELF_ENROLL',
  AGENT: 'AGENT',
} as const;